6-size.c

+++ 0x00-hello_world/6-size.c
@@ -8,10 +8,10 @@
 
 int main(void)
 {
-	printf("Size of a char: %lu byte(s)", sizeof(char));
-	printf("size of int: %lu byte(s)", sizeof(int));
-	printf("size of long int: %lu byte(s)", sizeof(long int));
-	printf("size of long int: %lu byte(s)", sizeof(long long int));
-	printf("size of float: %lu byte(s)", sizeof(float));
+	printf("Size of a char: %lu byte(s)", (unsigned long)sizeof(char));
+	printf("size of int: %lu byte(s)", (unsigned long)sizeof(int));
+	printf("size of long int: %lu byte(s)", (unsigned long)sizeof(long int));
+	printf("size of long int: %lu byte(s)", (unsigned long)sizeof(long long int));
+	printf("size of float: %lu byte(s)", (unsigned long)sizeof(float));
 	return (0);
 }
